Initial massaging of clientapi etc (not working yet)
